requirements.txt
docker_py==1.10.6
Flask==2.1.3
Requests==2.31.0
